+++ examples/src/bin/limitedclient.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 use std::io::{stdout, Read, Write};
 use std::net::TcpStream;
 
-use rustls::crypto::Ring;
+use rustls::crypto::ring::Ring;
 use rustls::OwnedTrustAnchor;
 
 fn main() {
