+++ rustls/src/ticketer.rs
@@ -1,5 +1,5 @@
 #[cfg(test)]
-use crate::crypto::Ring;
+use crate::crypto::ring::Ring;
 use crate::crypto::CryptoProvider;
 use crate::rand;
 use crate::server::ProducesTickets;
